pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "actuator-viz"
version = "0.1.0"
description = "Visualize and analyze multi-actuator control allocation systems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nathaniel Asiak", email = "asiaknathan@gmail.com" }
]
maintainers = [
    { name = "Nathaniel Asiak", email = "asiaknathan@gmail.com" }
]
keywords = [
    "robotics",
    "control-allocation",
    "actuators",
    "visualization",
    "ROV",
    "drone",
    "UAV",
    "UUV",
    "PX4",
    "ArduPilot",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "numpy>=1.21",
    "plotly>=5.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "typer>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
web = ["streamlit>=1.28"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
actuator-viz = "actuator_viz.cli:main"

[project.urls]
Homepage = "https://github.com/actuator-viz/actuator-viz"
Documentation = "https://github.com/actuator-viz/actuator-viz#readme"
Repository = "https://github.com/actuator-viz/actuator-viz"
Issues = "https://github.com/actuator-viz/actuator-viz/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/actuator_viz"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
